[sync.py] avoid infinite retry loop (#9858)
If a pod is unreachable for any reason, we previously retried forever. However,
a pods are ephemeral; we cannot assume they will return. Instead, if we fail to
contact a pod, we remove it from the pods list and log the error. If the pod
really does exist, the monitor_pods loop will attempt to initialize it again.