[query] Remove specialized from several usages with arrays (#9884)
* [query] Remove specialized from several usages with arrays

Scala calls functions that match on the type to avoid boxing. This is
slightly better than allocating, but so much worse than plain array
operations.

* [query] Remove specialized from several usages with arrays

Scala calls functions that match on the type to avoid boxing. This is
slightly better than allocating, but so much worse than plain array
operations.