[query-service] enable public use of query image
I screwed this up in the Shuffler PR. I need to remove the "image tag suffix" (the stuff after the final colon) before comparing
to a string like "gcr.io/hail-vdc/query".

I used the terminology from the Docker image specs:
- https://github.com/moby/moby/blob/master/image/spec/v1.md
- https://github.com/moby/moby/blob/master/image/spec/v1.2.md

But this terminology is kind of confusing. In "gcr.io/hail-vdc/query" the "repository" is "gcr.io/hail-vdc/query" and the tag is
not present.