[router-resolver][auth] fix latent bug in router-resolver (#9919)
* [router-resolver][auth] fix latent bug in router-resolver

The router resolver incorrectly assumed the contents of the `Authorization` header was a session
id. In fact, the structure of that header and X-Hail-Internal-Authorization is:

```
Bearer SESSION_ID
```

where `SESSION_ID` is a 44 base64 characters representing a 32 byte secret session id.

I also took this opportunity to centralize the parsing of bearer headers as
`gear.maybe_parse_bearer_header`.

---

This caused a failure because router-resolver, when checking that a user is properly authenticated,
would send:

```
Bearer Bearer SESSION_ID
```

which failed the 44-byte length check in auth/front_end.py.

* fix