[fs] add some parallelism to copy (#9922)
* [fs] add some parallelism to copy

This includes two changes:

First, add parallelism (async, single-core) to copy by parallelizing
over transfers, sources and files in directory sources.  This doesn't
get multi-core or intra-file parallelism yet.

copy() returns a CopyReport now.  It summarizes each transfer, source
and records counts of files copied, exceptions, etc.  To preserve the
old behavior, there is a flag that causes the first exception in the
report to be raised (if there is one) if there is one (this is the
default).

I will remark, this makes copy() very parallel and it will immediately
and do in parallel all the transfers, files and discovered files.

* delint