[address] retry with new IP as well
The IP address we got from `address` might be the IP of a pod that has been removed. When this happens
we get connection failed errors, which we treat as "transient". This change modifies the test to also
get a new IP address each time a transient error occurs. Not ideal, but more correct than previously.
There are forthcoming changes that I hope will more pervasively address this problem.