[query-service] use a UNIX Domain Socket for Py-Scala communication
The Py4J socket would crash regularly when I tried to execute several pipelines concurrently. I'm
fairly certain that Py4J is not thread-safe. With this change, I can execute many queries
concurrently.

Due to the removal of `statusForException`, it is much easier to read this diff when ignoring
whitespace changes.