[dev-docs] Hail For New Engineers (#10020)
* [dev-docs] Hail For New Engineers

This doc introduces the entire Hail project to new software engineers. This is a response to
Carolin's and Daniel's critiques of our on-boarding documentation.

I think there's a lot more to say, but I think it's time for some feedback before I continue.

The other things I want to document:
- How we use git with examples?
- What does genetic data look like?
- What is a GWAS?
- What is PCA as applied to genetics?
- A Tour of Hail Query (in the spirit of Soustroup's a Tour of C++).
- A Bibliography of Hail Team's recommended books.

* fix

* fix

* fixes

* fix