[query-service] get tests working (#10100)
* add fails_service_backend

* [hail] introduce and use mktemp and mktempd

We need temporary files and directories that clean themselves up and work across all the file
systems we support. `hl.mktemp` and `hl.mktempd` create such files and directories, respectively.

This was an unfortunately large change. I had to modify all our file systems to support `remove` and
`rmtree`. Since Google does not support directory creation, I had to do some groady `isinstance`
checks to ensure directories "exist" enough to be written into in all cases.

I also now know why my `/tmp` regularly fills up with weirdly named directories. I fixed Hail's
tests so they clean up after themselves in all cases.

* [hail] make is.hail.expr.ir.functions threadsafe

* [hail] fix use of row requiredness in lowerDistributedSort

* [query-service] handle void-typed IRs in query-service

* [query-service] make user cache thread-safe

* [query-service] bugfix: preserve globals through a shuffle

* [shuffler] log ShuffleCodecSpec anytime it is created

* [scala-lsm] bugfix: least key may equal greatest key

* [services] discovered new transient error

* [shuffler] more assertions in ShuffleClient

* [shuffler] bugfix: shuffler needs a HailContext to decode loci

* [query-service] move hail.jar earlier in Dockerfile

`hail.jar` is larger than the python code. If it appears first, then changes to
the Python code (which is small) does not create a fresh 100MB+ layer

* [query-service] permit pod scaling and remove cpu limit

* [query-service] simplify socket handling

* [batch] teach JVMJob where to find the hail configuration files

When `JVMJob` was added, it broke shuffler because the JVM processes live outside
of containers. The secrets all exist but in non-standard locations. We explicitly
configure extra environment variables for `JVMJob`s because only the Batch Worker
will know where the secrets exist.

* [query-service] switch to services team approved logging

Services team wants to know which class generated a log statement, so we use a fresh
logger for each class.

* [query-service] query workers need a hail context

* [query-service] use a UNIX Domain Socket for Py-Scala communication

The Py4J socket would crash regularly when I tried to execute several pipelines concurrently. I'm
fairly certain that Py4J is not thread-safe. With this change, I can execute many queries
concurrently.

Due to the removal of `statusForException`, it is much easier to read this diff when ignoring
whitespace changes.

* [query-service] run tests against query service

A few things need test resources in GCS.

Query tests are pretty standard. I use 8-way process parallelism with 3-way job parallelism for 24
concurrents tests. This will start-up slowly but once its running, each process uses practically no
CPU because query and batch are doing all the real work.

* [query-service] aiohttp.ClientSession must be created in async code

* [query-service][hail][build.yaml] address miscellaneous comments from cotton

* [query-service] use two containers sharing an empty volume

This makes tracking the logs from each process much simpler

* [query-service] in the service, log everything to stdout

Kubernetes expects logs to be sent to stdout/stderr

* [query-service] do not catch CancelledError

This error indicates that the coroutine was cancelled. When this happens, we need to
stop executing as soon as possible. The user has either disappeared (in which trying
to send them a response is pointless) or we have given up on the user (our pod is
being shutdown).

* [query-service] slightly more useful error message when socket dies

This includes the whole response object rather than an AssertionError with no further details.

* [shuffler] if we have an ExecuteContext, use it

Eventually, we need to always pass ShuffleClient an ExecuteContext. For now, let us remove
the allocation of a fresh context when we already have one on hand.

* [query-service] JSON Logging

* [query-service] pervasively retry all idempotent operations

* [hail] fix using

I got this wrong when I first wrote this function. We only want to catch exceptions from
`close`. Throwing the original exception inside this try block is kind of silly!

* [query] when testing, ensure our thread has an event loop

`asyncio.get_event_loop` refuses to start an event loop for you if you are not the
main thread. This is normally sensible: you should only have one event loop on one
thread and all async work should be sent to that thread's event loop. However,
during testing, it appears that pytest sometimes places us on a thread named
"Dummy-1". I do not understand why. I cannot find documentation about this. Anyway,
an easy fix is to create an event loop in whatever thread is running the tests.

* [query][lir] avoid dumping massive classes onto stderr

Dumping a huge class file to stderr is bad in most cases, but particularly bad in the service.
Instead, I slurp the class file into a string and put that into the logging system.

* [query] do not print misleading log in RegionPool.finalizer

* [query-service] better logging information

* [query-service] remove race conditions in user management

removeUser can race with addUser and is unused anyway. Two threads can
try to addUser at the same time.

* [lir] make LIR genName thread-safe

* [query-service] send a token for job identifiability

This token *is not for restarts or dropped connections*. This token just helps
identify the logs for a given job.

* [query-service] reduce image size by ~2GB

* [query-service] retry CLOSE, CLOSED (i.e. connection dropped)

Exponentially back off on connection drop. Ideally, the service will record
intermediate states of execution and we can resume a query. Currently, this is
not possible. Instead, we simply restart. This should allow tests to work, but
needs to be addressed more seriously.

* [query-service] remove old test code

* [query-service] all operations are idempotent

* [make] fix config.mk

* [devbin] teach devbin/functions.py about multiple containers

* [batch] remove batch_worker_image false dependency on service_base_image

* [java-services] avoid NPEs in isTransientError

* [dependencies] upgrade gcsfs to 0.7.2 to fix GoogleFS rmtree issue

gcsfs 0.2.2 is quite old and has a number of bugs. In particular, it cannot remove multiple GCS
objects in parallel, so it fails on `rmtree`.

* [query-service] enable auto-scaling for PR and dev deploy

* [query-service] retry EndOfStream errors from java

We only receive an EndOfStream when the Java process has violated our
socket protocol, or we have an implementation bug. Assuming implementation bugs
are not released into production, we want to retry in this case.

* [batch] remove restrictions on PR and dev batch pools

The query service needs a lot of computation to test itself. This lets us scale up to
128 full-sized cores. The query tests can keep that saturated for a while.

* [query-service] resolve last two issues

GCS stat apparently does not always include the path. Also, this phsyical_key_truncation seems
to fail in the service, consistently due to the context file being unexpectedly short.

* fixup config.mk

* no braces

* remove races from add_user

* use __name__

* namespace the query dict to a user

* cleanup user queries

* remove n=32 in secret_alnum_string

* remove hail query bucket

* address sockets.py issues

* remove pylintrc from service-java-run-base

* remove commented line in query/log4j.properties

* fix pylint

* bug fixes

* skip test_pedigree in service because it is flaky

* remove duplicate query-tests tls

* fix

* fix skip