[query-service] concurrently use multiple Hail Query JAR versions
The high-level problem:

The Hail Query Service is redeployed with each commit to `main`. Each deployment has a new JAR file
whose ABI is backwards-incompatible.

The high-level solution:

Hail Batch Workers can load the JAR for a given Hail version on-demand. A fresh class loader for
each Hail version allows the classes to co-exist in the same JVM. We cache jars on the local
filesystem.

---

`javac` compiles Java files to JVM Bytecode. JVM Bytecode is normally stored in `class` files. A JAR
file is, essentially, a TAR file of a directory of class files.

`java` needs to find the `class` file that defines any Class. A `ClassLoader` defines:

1. (`findClass`) How to *find* the definition of a Class known to the current `ClassLoader`.

2. (`findResource`) How to *find* an arbitrary file known to the current `ClassLoader`.

3. (`loadClass` and `getResource`) The order in which to find a class in a set of
  `ClassLoader`s (e.g. if two `ClassLoader`s know about the same Class, which one should load the
  class?)

Every `ClassLoader` has a `parent` `ClassLoader`. The default implementation of `loadClass` and
`getResource` prefers loading classes from its parent ClassLoader before anything else. We invert
the loading order to allow multiple definitions of the same Class in the same JVM. In particular,
each instance of `LoadSelfFirstURLClassLoader` prefers to use its own definition of a Class. Each
`LoadSelfFirstURLClassLoader` instance knows about one version of the Hail JAR.

The remaining subtle issue is how to load resources. For example, `HailBuildInfo` needs to load the
build info resource file. To do so, you need an instance of a `ClassLoader` that can find the
file you want. Often times, you use `this.getClass().getClassLoader()`, which is the class loader
used to load the current class. Hail does not do this. I believe we do not do this because of issues
with how TestNG loads classes. :sigh: As a result, I also modify the worker Thread's
ContextClassLoader for the duration of the execution of an alternative version of Hail.

---

I've already updated the cluster with the new bucket and the corresponding change to the
`global-config` secret. We'll should notify Leo et al. about the Terraform changes.

---

Small changes and fixes:

- Rename `key.json` to `/worker-key.json` for clarity, this is the worker's own GCP service account
  key.
- Create a test query-gsa-key in test and dev namespaces.
- Add terraform rules for the query service account. It already existed, but it was missing from the
  Terraform file. You can verify the permissions grant by inspecting `gsutil iam get
  gs://hail-query`.
- The `query` user was missing from bootstrap-create-accounts.
- `hail-ubuntu-stmp` was missing from `docker/Makefile`'s `clean` rule
- Use a dummy `WorkerBackend` when we're on the worker. The worker isn't allowed to call these
  methods anyway, that would amount to Hail Queries inside of Hail Queries, madness!
- New transient error in Java.