[aiotools] add better debugging information for flaky test (#10243)
This test fails often ([for example](https://ci.hail.is/batches/234836/jobs/73)). I do not believe the test itself is being cancelled
by pytest-asyncio. Instead, I believe the aiotools code inadvertantly allows a
`CancelledError` to bubble to the top-level. Unfortunately, `asyncio` appears
to treat that as if the *task itself was cancelled*. Because it treats the
task as cancelled, we lose the stack trace (which would tell us where
we were just before the cancellation occurred).