[batch] make JVM logs real-time
This is a simplified implementation of `Process.communicate`. We feed lines into the log one-by-one
until we reach the end of both stdout and stderr. When both stdout and stderr have been closed by
the child process, we wait for the process to exit. At any point in time, the most recent log is
available to us.