[website] fix for issues raised in #10277
@nawatts points out that the docs are missing the font-awesome icons. This new website
stuff is confusing! Here is how to think about it going forward:

This PR is the *long-term fix*. It will not resolve the current production issue. The easiest fix
for that is to release a new version of Hail, which will generate a new version of the docs. That
new version will be compatible with the currently released `website`.

We have two phases: docs-generation-time and serve-time. Importantly, we always need to support the
*previously released documentation* with the current `main` `website`.

This PR makes things work like this:

At docs-generation-time (which happesn every time we release a new version of the Hail package to
PyPI), Sphinx uses `dynamic-base.html` to build the docs website. Those HTML files will look like
this:

```
<!DOCTYPE html>
<html lang="en">
  <head>
    <title>Hail | {% block title %}{% endblock %}</title>

    {% include "base-head.html" %}

    <!-- Sphinx will insert some header stuff here -->
  </head>
  <body>
    {% include "nav-top.html" %}
    <div id="main">
      {% raw %}
      <!-- Sphinx will insert the actual documentation here -->
      {% endraw %}
    </div>
    {% include "nav-bottom.html" %}
    {% include "base-foot.html" %}
  </body>
</html>
```

Note that this is *still a Jinja2 template!*

At serve-time, `website` will run this through Jinja2 templating *a second time*. At this time,
we'll use the latest `base-head.html`.

In particular, suppose I need access to a new CSS file in `nav-top.html`. I can modify
`base-head.html`. Since `base-head.html` is added to the web page *at serve-time*, it will include
all the latest changes.

In contrast, everything inside the `#main` `div` *only changes when Hail is released* because it is
only updated at docs-generation-time.