[hailtop][batch] Fix DockerHub
Tags and digests have no affect on whether an image is one of the hailgenetics
images that are stored in both GCR and DockerHub. This change ignores the tag and
digest when checking if we should warn about using DockerHub.

In doing this, I consolidated and made more robust our Docker-image-reference parsing.