[utils] make sleep_and_backoff a bit more predictable
Multiplying by random number from 0 to 1 uniformly seems a bit too aggressive and harder for us, as
the programmers, to understand.

Consider that, one in a hundred uses of this function, we will get a sequence of random draws with
each number less than 0.1:

    0.1, ..., 0.05, ..., 0.09

For that same sequence of draws, the sequence of delays, if started at 0.1, is this:

    0.1, 0.2, 0.4, 0.8, 1.6, 3.2, 6.4, 12.8, 25.6, 51.2

The next delay will be 60. Suppose the next draw is 1.0. We will have waited no more than ~5 seconds
[1], but our next delay is 60 seconds!

This PR proposes using

    t = delay * random.uniform(0.9, 1.1)

which ensures that each delay is at least 90% of the expected delay. In the worst case scenario, by
the 11th attempt we have waited, in total, at least ~45 seconds.

I think this is better because whatever code we run between the delays is highly likely to follow an
exponential back-off up to 30s rather. Currently, there is a one-in-one-hundred chance that we have
ten delays of <2.5 seconds followed by something relatively large like 30s.