[query-service] concurrently use multiple Hail Query JAR versions  (#10390)
* [query-service] concurrently use multiple Hail Query JAR versions

The high-level problem:

The Hail Query Service is redeployed with each commit to `main`. Each deployment has a new JAR file
whose ABI is backwards-incompatible.

The high-level solution:

Hail Batch Workers can load the JAR for a given Hail version on-demand. Although not a long-term
solution, we currently start a fresh JVM for each job. As a result, we can simply start the JVM with
the correct JAR on its classpath. We cache jars on the local filesystem.

I had to abandon the old approach for two reasons:

1. Multiple JVMs race to download the JAR. In the new approach, the python worker process uses a
   lock to ensure at most one coroutine is downloading a given version of a JAR at the same time.

2. The JVM assumes that a child ClassLoader does not redefine a class from the parent
   ClassLoader. That's why ClassLoaders always prefer to load a class from the parent ClassLoader's
   classes.

When we decide to re-use JVMs or use a single multi-threaded JVM, we'll need to ensure the top-level
ClassLoader *does not have Hail on its classpath*. I looked briefly at this approach and found it
more work than the current approach.

---

My apologies for eliminating JVMProcess in this PR. It's an unrelated change which facilitated my
understanding worker.py. I essentially inlined JVMProcess into JVMJob and eliminated any duplicative
code.

---

After making this change I restored the tests. Some tests had bitrotted. In the process of fixing
those tests, I found a few other bugs. Fixing these lower-level bugs unlocked a number of new
tests.

A couple tests (which were added since the service tests were removed) had to be marked as
failing.

Here are the bugs I fixed:

1. Correct the error message raised when tests are run in a non-main thread (we look for this
   message and start an event loop for Hail's async code because asyncio refuses to start an event
   loop in a non-main thread).

2. Use a `SafeRow` to copy the globals data out of a Region and into durable, GC'ed objects.

3. Re-enable serialization of GoogleStorageFS (including its private key, which we really shouldn't
   do; Tim is working on it), which was broken (presumably) when we changed Scala versions. The
   `var` modifier ensures the name is compiled as a JVM field.

4. Correctly convert from a `Byte` to an `Int`. By default `Byte` to `Int` conversion (which is done
   automatically when you return a `Byte` from a function whose return type is `Int`) is
   sign-preserving. That means that the byte `0000 1111` is converted to the `Int` 15 and the byte
   `1000 1111` is converted to the `Int` -113. The contract of
   [`InputStream.read`](https://docs.oracle.com/javase/8/docs/api/java/io/InputStream.html#read--)
   is to return the unsigned integeral value of the next `Byte` or `-1` if we've reached the end of
   the stream. `DataInputStream` treats any negative value as EOS which lead to perplexing EOSes
   when reading data from GCS.

5. Retain the `gs://` protocol when reading MTs and Ts. `uriPath` strips *all* protocols. Before the
   Query Service, these code paths were only used by the LocalBackend. In the LocalBackend, the only
   URIs generated are `file://`. However, UNIX/JVM file system operations do not support URIs, they
   want bare paths.

6. Implement missing cases for EShuffle and PShuffle.

7. BLAS and LAPACK need to be thread-local.

8. The LSM-tree used by the Shuffler needs to permit multiple values for the same key. There was
   some subtlety here around fine-grained locking. Unfortunately, the LSM doesn't expose the right
   operations (putIfAbsent) to make this easy, so I had to use a concurrent hash map of locks.

---

I've already updated the cluster with the new bucket and the corresponding change to the
`global-config` secret. We'll should notify Leo et al. about the Terraform changes.

---

Small changes and fixes:

- Rename `key.json` to `/worker-key.json` for clarity, this is the worker's own GCP service account
  key.

- Create a test query-gsa-key in test and dev namespaces.

- Add terraform rules for the query service account. It already existed, but it was missing from the
  Terraform file. You can verify the permissions grant by inspecting `gsutil iam get
  gs://hail-query`.

- The `query` user was missing from bootstrap-create-accounts.

- Remove the auto-scaling and remove the k8s grace period. Neither of these really works right. We
  will live without the auto-scaling for now. I have to fix the grace period thing before we let
  users run real pipelines.

* shuffle is non-deterministically failing

* fix

* fix

* cleanup timing stuff

* fix grammar in fails comment

* use one envvar

* note about backends

* use val instead of var

* fix

* order by does not error anymore

* fix

* fix

* new transient error

* fix