[query] Fix bad memory leak triggered by set contains IR function
CHANGELOG: Fixed a memory leak triggered by `hl.literal(...).contains(...)

This bug is present elsewhere in the code generator, but the set
contains function is probably the worst place for it to happen.
This leads to a full copy of the set where the binary search
is executed.

The core problem was a bug in PArrayBackedContainer not casting
its codes properly, leading to the no-op coerce logic in PCanonicalArray
being bypassed in favor of the generic copy-the-world implementation of
`store`. The test I have added catches the memory leak, but now fails at
compile time at the assertion to `PCanonicalArray.store` instead.