[query] Fix table type of MT cols subfolder
This change makes the `test_read_stored_cols` work via lowering.
I'm still unsure if we should support the cols/rows/entries subfolders
of a MT being valid tables themselves, but certainly it's incorrect
to describe the written cols table as having the same key as the MT
columns when it's unordered.

This change strips the col key off the tabletype generated in the
file, to reflect the ordering status.