[compiler] Add SUnreachable SType for NA / Die (#10539)
* [compiler] Add SUnreachable SType for NA / Die

While only used in a few places, this helps us generate
much better code in the case where we emit error-checking
IRs as below:

```
If
  <error_condition>
  Die err_msg
  <value we want>
```

The code generator for the `If` node uses `SType.canonical` to
choose its result type, and casts both consequent and alternate
values to that type. We want the stype of the result here to be
the type of the alternate `<value we want>`, which we can achieve
by adding unreachable types/codes for the `SType.canonical` logic.

* remove unnecessary narrow cast in PCanonicalArray.store

* fix binary

* fix string ord

* more fixes

* oops

* fix locus

* fixes

* oops

* extend settable

* fix rebase

* fix createColumnMajorCode

* fix

* fix matmul

* fix ebinary encoder

* fixes

* fix

* fix

* fix fieldidx cast

* fix compilation

* add inner requiredness to PType.canonical

* fix rebase

* fix rebase

* I did make mistakes

* address comments

* fix rebase