[compiler] Rework reference genome code setup
Reference Genomes now manage liftover and sequences with serializable
metadata fields. The actual fasta reader config and liftover map are
transient. They are set by a new method called `heal` that synchronizes
liftover and sequences with the file paths now present in the class.
This removes the need for codeSetup and similar methods.