[compiler] Lazily initialize reference genomes in generated code  (#10523)
* Add liftover test to catch RG initialization bug

* [compiler] Rework reference genome code setup

Reference Genomes now manage liftover and sequences with serializable
metadata fields. The actual fasta reader config and liftover map are
transient. They are set by a new method called `heal` that synchronizes
liftover and sequences with the file paths now present in the class.
This removes the need for codeSetup and similar methods.

* fix tests

* fix heal when no sequence has been added

* Don't use broadcasted reference genomes in resultWithIndex

* heal doesn't check chain files either

The global reference genomes are not usually initialized on workers, so,
in keeping in similar form to the old addLiftoverFromFS and
addSequenceFromReader, heal doesn't check validity of the chain files or
sequence files. The responsibility there is with addSequence and
addLiftover.

* check if liftover and fasta config actually need to be updated in heal

* initialize liftoverMap in heal

* hasSequence and hasLiftover based on fastaFilePath and chainFiles

* fix removeSequence

Co-authored-by: Tim Poterba <tpoterba@gmail.com>