[asyncfs] fix deadlock in S3 backend (#10704)
* boto is synchronous.  The AsyncFS create interface returns a writable
file-like object, which means the S3 request has to be run in a thread
in the background.  Originally, I was using the thread pool for that.
However, if we fill the thread pool with create requests, then there
are no threads for writes, so the code can deadlock.  Instead, run the
create requests in a new thread.

* fix exception propagation