[hailtop] add hailtop.aiotools.copy (#10778)
* [hailtop] add hailtop.aiotools.copy

The following operation copies every file in gs://danking/test/ into a folder named
s3://hail-test-bucket/test/.

   python3 -m hailtop.aiotools.copy null '[{"from": "gs://danking/test", "to": "s3://hail-test-bucket/test"}]'

The `null` can be replaced with a quoted GCP project id if one of the sources is a Requester Pays bucket. For example:

   python3 -m hailtop.aiotools.copy '"broad-ctsa"' '[{"from": "gs://hail-datasets-us/", "to": "s3://hail-datasets-us-east-1"}]'

As you can see, the syntax is rough, but expressive. Explicitly listing all the files to be
transferred should not impair the throughput.

   python3 -m hailtop.aiotools.copy null '[
       {"from": "gs://danking/test/data1", "to": "s3://hail-test-bucket/test/data1"},
       {"from": "gs://danking/test/data2", "to": "s3://hail-test-bucket/test/data2"},
       {"from": "gs://danking/test/sub-directory1", "to": "s3://hail-test-bucket/test/sub-directory1"},
       ...
   ]'

Better syntax is forthcoming: https://github.com/hail-is/hail/pull/9913

* also remove batch.copy

* address pylint

* align messages

* fix missing scheme case

Co-authored-by: Cotton Seed <cotton@alum.mit.edu>