[query/compiler] Harmonize FS.open behavior in python (#10812)
hadoop_open has a somewhat strange behavior, when the global fs is
HadoopFS, BGzip and Gzip files are handled transparently by file
extension, so python reads and writes uncompressed data. This is not the
case if the global fs is LocalFS or GoogleCloudStorageFS. We'd
eventually like to move away from this behavior for HadoopFS altogether,
but we cannot change the behavior of hadoop_open without breaking user
code.

To that end, rewrite HadoopFS.open to ignore codecs, and add legacy_open
to preserve the old behavior.

As a result of this, we also implement the seekable interface in python
for HadoopFS openend files.