[compiler] Add Apache avro file reading (#10793)
* [compiler] Add Apache avro import

The support is very limited at this time, only basic data types,
["null", type] unions, and record (struct) types.

The intent is that for every field in the record, generate the
appropriate code for a stack struct.

* Almost complete implementation

* rowRequiredness for AvroPartitionReader

* Add simple avro partition test

* support both ["null", type] and [type, "null"] unions

* AvroPartitionReader fixes

- Fixup code generation for AvroPartitionReader
- Fix tests

* Set static helper methods private

* add avro table reader

* python avro table reader wip

* AvroTableReader json (de)serialization

* AvroTableReader can be lowered efficiently

* AvroTableReader.lower

* fix lower

* exclude avro jackson dependency from bundle

* Make HadoopSeekablePyReader more pythonic

It now implements the full seek options for python, in order to support
SEEK_END, also get the FileStatus for the size when opening the stream.

* context manager cleanup

* hail requires avro

* add simple import_avro test

* fix python AvroTableReader equality

* Use EmitCode.fromI over EmitCode.apply

* Use a literal for paths

* Add test for smaller requested type

* Just call it AvroSchemaSerializer

* Add unsafe avro table reader options

This allows a user to decleare a key and a partitioning. They must be
correct with respect to the data, and no checks are done to ensure that
they are.