[batch] make tests resilient to concurrent batches
These tests only work if fewer than 200 batches were created between the batches
under test and the `list_batches` API call. This is obviously not necessarily
true in general and *certainly* not true when the Hail Query on Hail Batch is
spewing hundreds of batches per second into the system.