[query] break up huge tests
These two tests take a very long time in the service and do not benefit from the
massive horizontal scalability of Hail Query on Hail Batch. These tests also
somewhat frequently cause OOMs in the service backend, so the point is perhaps
moot. Nonetheless, I think this is an overall positive change.