[batch] close file store *after* work stops (#10949)
Currently there is a race condition in which we can shut down the filestore before
the aiohttp app stops accepting connections. If that happens, jobs will get
partially scheduled on the worker, but there will be no working file store so the
jobs cannot complete successfully. The partially scheduled jobs in turn leave the
worker in a bad, non-idle state.