Switch to a log(n_samples, branch_factor) binning strategy for vds merging
Instead of keeping a sorted list, the VDS arguments are kept in bins
keyed by floor(log(n_samples, branch_factor)). The new VDS merge
candidates are chosen from the front of the smallest bin and from the
back of larger bins if the smallest bin has fewer than branch_factor
elements. This keeps the ouput order more stable with respect the order
of the input GVCFs and the sorted (by number of samples) VDS arguments.